refactor(ofm_core/contracts): extract free-agent cluster
Move the free-agent signing path out of contracts/mod.rs into
contracts/free_agent.rs: project_free_agent_contract_impact (affordability preview),
offer_free_agent_contract (the signing flow) and free_agent_signed_message.
Byte-faithful move via script; re-exported with `pub use` so the two public entry
points stay resolvable as `ofm_core::contracts::*`.

387 ofm_core lib + 28 contracts integration tests pass; clippy clean in contracts/.
